Bd fitting

+++ galfind/Catalogue.py
@@ -279,16 +279,28 @@ def jaguar_phot_labels(
     err_labels = {aper_diam * aper_diams.unit: [] for aper_diam in aper_diams.value}
     return phot_labels, err_labels
 
-def scattered_phot_labels(
+def scattered_phot_labels_inst(
     filterset: Multiple_Filter, 
     aper_diams: u.Quantity, 
     **kwargs
 ) -> Tuple[Dict[str, str], Dict[str, str]]:
     assert "min_flux_pc_err" in kwargs.keys(), \
         galfind_logger.critical("min_flux_pc_err not in kwargs!")
-    phot_labels = {aper_diam * aper_diams.unit: [f"{filt_name}_scattered" for filt_name in filterset.band_names] for aper_diam in aper_diams.value}
-    err_labels = {aper_diam * aper_diams.unit: [f"{filt_name}_err" for filt_name in filterset.band_names] for aper_diam in aper_diams.value}
+    phot_labels = {aper_diam * aper_diams.unit: [f"{filt.instrument_name}.{filt.band_name}_scattered" for filt in filterset] for aper_diam in aper_diams.value}
+    err_labels = {aper_diam * aper_diams.unit: [f"{filt.instrument_name}.{filt.band_name}_err" for filt in filterset] for aper_diam in aper_diams.value}
     return phot_labels, err_labels
+
+def scattered_phot_labels(
+    filterset: Multiple_Filter,
+    aper_diams: u.Quantity,
+    **kwargs
+) -> Tuple[Dict[str, str], Dict[str, str]]:
+    assert "min_flux_pc_err" in kwargs.keys(), \
+        galfind_logger.critical("min_flux_pc_err not in kwargs!")
+    phot_labels = {aper_diam * aper_diams.unit: [f"{filt.band_name}_scattered" for filt in filterset] for aper_diam in aper_diams.value}
+    err_labels = {aper_diam * aper_diams.unit: [f"{filt.band_name}_err" for filt in filterset] for aper_diam in aper_diams.value}
+    return phot_labels, err_labels
+
 
 def galfind_mask_labels(
     filterset: Multiple_Filter, 
@@ -304,6 +316,21 @@ def galfind_depth_labels(
     return {aper_diam * aper_diams.unit: [f"loc_depth_{filt_name}" \
         for filt_name in filterset.band_names] \
         for aper_diam in aper_diams.value}
+
+def scattered_depth_labels(
+    filterset: Multiple_Filter,
+    aper_diams: u.Quantity,
+    **kwargs
+) -> Dict[str, str]:
+    return {
+        aper_diam * aper_diams.unit: 
+        [
+            f"loc_depth_{filt.instrument_name}.{filt.band_name}"
+            for filt in filterset
+        ]
+        for aper_diam in aper_diams.value
+    }
+
 
 def load_bool_Table(
     tab: Table,
@@ -1000,31 +1027,34 @@ class Catalogue(Catalogue_Base):
                 "KRON_RADIUS", 
                 "sex_KRON_RADIUS", 
             )
-            A_image_arr = self.load_band_properties_from_cat(
-                "A_IMAGE",
-                "sex_A_IMAGE",
-                update = False
-            )
-            [setattr(gal, "sex_A_IMAGE", A_image[self.data[0].filt_name]) for gal, A_image in zip(self, A_image_arr)]
-            A_image_as_arr = [{band_data.filt_name: kron_radius[band_data.filt_name] * A_image[band_data.filt_name] * band_data.pix_scale \
-                for band_data in self.data} for kron_radius, A_image in zip(kron_radii, A_image_arr)]
-            [gal.load_property(A_image_as, "sex_A_IMAGE_AS") for gal, A_image_as in zip(self, A_image_as_arr)]
-            B_image_arr = self.load_band_properties_from_cat(
-                "B_IMAGE",
-                "sex_B_IMAGE",
-                update = False
-            )
-            [setattr(gal, "sex_B_IMAGE", B_image[self.data[0].filt_name]) for gal, B_image in zip(self, B_image_arr)]
-            B_image_as_arr = [{band_data.filt_name: kron_radius[band_data.filt_name] * B_image[band_data.filt_name] * band_data.pix_scale \
-                for band_data in self.data} for kron_radius, B_image in zip(kron_radii, B_image_arr)]
-            [gal.load_property(B_image_as, "sex_B_IMAGE_AS") for gal, B_image_as in zip(self, B_image_as_arr)]
-            theta_image_arr = self.load_band_properties_from_cat(
-                "THETA_IMAGE",
-                "sex_THETA_IMAGE",
-                multiply_factor = u.deg,
-                update = False,
-            )
-            [setattr(gal, "sex_THETA_IMAGE", theta_image[self.data[0].filt_name]) for gal, theta_image in zip(self, theta_image_arr)]
+            try:
+                A_image_arr = self.load_band_properties_from_cat(
+                    "A_IMAGE",
+                    "sex_A_IMAGE",
+                    update = False
+                )
+                [setattr(gal, "sex_A_IMAGE", A_image[self.data[0].filt_name]) for gal, A_image in zip(self, A_image_arr)]
+                A_image_as_arr = [{band_data.filt_name: kron_radius[band_data.filt_name] * A_image[band_data.filt_name] * band_data.pix_scale \
+                    for band_data in self.data} for kron_radius, A_image in zip(kron_radii, A_image_arr)]
+                [gal.load_property(A_image_as, "sex_A_IMAGE_AS") for gal, A_image_as in zip(self, A_image_as_arr)]
+                B_image_arr = self.load_band_properties_from_cat(
+                    "B_IMAGE",
+                    "sex_B_IMAGE",
+                    update = False,
+                )
+                [setattr(gal, "sex_B_IMAGE", B_image[self.data[0].filt_name]) for gal, B_image in zip(self, B_image_arr)]
+                B_image_as_arr = [{band_data.filt_name: kron_radius[band_data.filt_name] * B_image[band_data.filt_name] * band_data.pix_scale \
+                    for band_data in self.data} for kron_radius, B_image in zip(kron_radii, B_image_arr)]
+                [gal.load_property(B_image_as, "sex_B_IMAGE_AS") for gal, B_image_as in zip(self, B_image_as_arr)]
+                theta_image_arr = self.load_band_properties_from_cat(
+                    "THETA_IMAGE",
+                    "sex_THETA_IMAGE",
+                    multiply_factor = u.deg,
+                    update = False,
+                )
+                [setattr(gal, "sex_THETA_IMAGE", theta_image[self.data[0].filt_name]) for gal, theta_image in zip(self, theta_image_arr)]
+            except:
+                pass
             
             for gal in self:
                 for aper_diam in gal.aper_phot.keys():
@@ -1306,6 +1336,7 @@ class Catalogue(Catalogue_Base):
         plot_lowz: bool = True,
         wav_unit: u.Unit = u.um,
         flux_unit: u.Unit = u.ABmag,
+        log_fluxes: bool = False,
         crop_name: Optional[str] = None,
         collate_dir: Optional[str] = None,
         imshow_kwargs: Dict[str, Any] = {},
@@ -1350,6 +1381,7 @@ class Catalogue(Catalogue_Base):
                 n_cutout_rows = n_cutout_rows,
                 wav_unit = wav_unit,
                 flux_unit = flux_unit,
+                log_fluxes = log_fluxes,
                 imshow_kwargs = imshow_kwargs,
                 norm_kwargs = norm_kwargs,
                 aper_kwargs = aper_kwargs,
@@ -1387,14 +1419,15 @@ class Catalogue(Catalogue_Base):
                     os.remove(symlink_path)
             # create symlink to selection folder for diagnostic plots
             for gal, out_path in zip(self, out_paths):
-                selection_path = f"{collate_dir}/{str(gal.ID)}.png"
-                funcs.make_dirs(selection_path)
-                try:
-                    os.symlink(out_path, selection_path)
-                except FileExistsError:  # replace existing file
-                    if Path(out_path).is_file():
-                        os.remove(selection_path)
+                if out_path is not None:
+                    selection_path = f"{collate_dir}/{str(gal.ID)}.png"
+                    funcs.make_dirs(selection_path)
+                    try:
                         os.symlink(out_path, selection_path)
+                    except FileExistsError:  # replace existing file
+                        if Path(out_path).is_file():
+                            os.remove(selection_path)
+                            os.symlink(out_path, selection_path)
 
     # Number Density Function (e.g. UVLF and mass functions) methods
 
@@ -1404,6 +1437,7 @@ class Catalogue(Catalogue_Base):
         aper_diam: u.Quantity,
         SED_fit_code: SED_code,
         z_step: float = 0.01,
+        unmasked_area: Union[str, List[str], u.Quantity] = "selection",
     ) -> NDArray[float]:
         assert hasattr(self, "data"), \
             galfind_logger.critical(
@@ -1415,6 +1449,7 @@ class Catalogue(Catalogue_Base):
             aper_diam = aper_diam,
             SED_fit_code = SED_fit_code,
             z_step = z_step,
+            unmasked_area = unmasked_area,
         )
     
 
@@ -1426,17 +1461,21 @@ class Catalogue(Catalogue_Base):
         aper_diam: u.Quantity,
         mode: str,
         depth_region: str = "all",
+        min_flux_pc_err: float = 10.0,
     ):
         assert all(aper_diam in gal.aper_phot.keys() for gal in self)
         # load galaxy depths from the average depths of the field
-        self._update_depths_from_data(aper_diam, mode, depth_region)
+        if hasattr(self, "data"):
+            self._update_depths_from_data(aper_diam, mode, depth_region)
         # calculate photometric errors from these newly inserted depths
-        self._update_errs_from_depths(aper_diam)
+        self._update_errs_from_depths(aper_diam, apply_min_flux_pc_err = False)
         # scatter each set of fluxes once by the calculated errors
         [
             gal.aper_phot[aper_diam].scatter_fluxes(update = True) 
             for gal in tqdm(self, desc = "Scattering catalogue fluxes", total = len(self))
         ]
+
+        self._update_errs_from_depths(aper_diam)
 
     def _update_depths_from_data(
         self: Self,
@@ -1462,16 +1501,21 @@ class Catalogue(Catalogue_Base):
         self: Self,
         aper_diam: u.Quantity,
         default_min_flux_pc_err: float = 10.0,
+        apply_min_flux_pc_err: bool = True,
     ) -> NoReturn:
-        if "min_flux_pc_err" in self.cat_creator.load_phot_kwargs.keys():
-            min_flux_pc_err = self.cat_creator.load_phot_kwargs["min_flux_pc_err"]
+        if apply_min_flux_pc_err:
+        
+            if "min_flux_pc_err" in self.cat_creator.load_phot_kwargs.keys():
+                min_flux_pc_err = self.cat_creator.load_phot_kwargs["min_flux_pc_err"]
+            else:
+                galfind_logger.warning(
+                    f"No 'min_flux_pc_err' in {self.cat_creator.load_phot_kwargs.keys()=}." + \
+                    f" Using {default_min_flux_pc_err=}!"
+                )
+                min_flux_pc_err = default_min_flux_pc_err
         else:
-            galfind_logger.warning(
-                f"No 'min_flux_pc_err' in {self.cat_creator.load_phot_kwargs.keys()=}." + \
-                f" Using {default_min_flux_pc_err=}!"
-            )
-            min_flux_pc_err = default_min_flux_pc_err
-        [
-            gal.aper_phot[aper_diam]._update_errs_from_depths(min_flux_pc_err)
+            min_flux_pc_err = 0.
+
+        [gal.aper_phot[aper_diam]._update_errs_from_depths(min_flux_pc_err)
             for gal in tqdm(self, desc = "Updating catalogue errors from average depths", total = len(self))
         ]
